Move classes from platform.common to generator.common
Moved GeneratorSuite and CommendsProcessor from
"...platform.common" package to "...generator.common". Before
the generator refactoring project started, generator-related
classes were scattered inconsistently between these "platform"
and "generator" packages. The refactoring consolidates
everything under "generator", making "platform" obsolete.
