+++ gluecodium/src/main/java/com/here/gluecodium/Gluecodium.kt
@@ -26,6 +26,7 @@ import com.here.gluecodium.cli.OptionReader
 import com.here.gluecodium.cli.OptionReaderException
 import com.here.gluecodium.common.LimeLogger
 import com.here.gluecodium.generator.common.GeneratedFile
+import com.here.gluecodium.generator.common.GeneratorSuite
 import com.here.gluecodium.generator.common.templates.TemplateEngine
 import com.here.gluecodium.generator.java.JavaGeneratorSuite
 import com.here.gluecodium.loader.getLoader
@@ -33,7 +34,6 @@ import com.here.gluecodium.model.lime.LimeModel
 import com.here.gluecodium.model.lime.LimeModelLoader
 import com.here.gluecodium.model.lime.LimeModelLoaderException
 import com.here.gluecodium.output.FileOutput
-import com.here.gluecodium.platform.common.GeneratorSuite
 import com.here.gluecodium.validator.LimeDeprecationsValidator
 import com.here.gluecodium.validator.LimeEnumeratorRefsValidator
 import com.here.gluecodium.validator.LimeExternalTypesValidator
