+++ gluecodium/src/main/java/com/here/gluecodium/generator/swift/SwiftGenerator.kt
@@ -20,6 +20,7 @@
 package com.here.gluecodium.generator.swift
 
 import com.here.gluecodium.generator.cbridge.CBridgeCollectionNameResolver
+import com.here.gluecodium.generator.common.GeneratorSuite
 import com.here.gluecodium.generator.common.modelbuilder.LimeTreeWalker
 import com.here.gluecodium.model.lime.LimeElement
 import com.here.gluecodium.model.lime.LimeNamedElement
@@ -29,7 +30,6 @@ import com.here.gluecodium.model.swift.SwiftFile
 import com.here.gluecodium.model.swift.SwiftModelElement
 import com.here.gluecodium.model.swift.SwiftStruct
 import com.here.gluecodium.model.swift.SwiftTypeDef
-import com.here.gluecodium.platform.common.GeneratorSuite
 
 class SwiftGenerator(
     private val limeReferenceMap: Map<String, LimeElement>,
